+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Bugfix: exceptions raised from a property are no longer swallowed when
   getting a key from a context stack (issue #110).
 * Bugfix: lambda section values can now return non-ascii, non-unicode strings (issue #118).
+* More robust handling of byte strings in Python 3.
 
 0.5.2 (2012-05-03)
 ------------------